Improve Performance by Removing all Unique Checks
The unique checks are not necessary because the identifiers (UUID) of
the entities are unique anyway so that Datomic throwns an exception. The
exception-handler branch will improve the error messages. But for now
the excact shape of error messages is not important.

I also use direct lookups of study-events and subjects by there
identifier in the upsert case which should be much faster than iterating
over the collections. This might even the most serious performance
winner here.